requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiohttp>=3.9.1
requests>=2.31.0
msal>=1.25.0
pyotp>=2.9.0
cryptography>=42.0.2
python-jose>=3.3.0
pyyaml>=6.0.1
qrcode>=7.4.2
psutil>=5.9.8
Flask>=3.0.2 